+++ crates/bevy_ecs/src/system/mod.rs
@@ -112,6 +112,8 @@ mod query;
 mod system;
 mod system_param;
 
+use std::borrow::Cow;
+
 pub use combinator::*;
 pub use commands::*;
 pub use exclusive_function_system::*;
@@ -186,6 +188,51 @@ where
     S::System: ReadOnlySystem,
 {
     assert_is_system(system);
+}
+
+/// Conversion trait to turn something into a [`System`].
+///
+/// Use this to get a system from a function. Also note that every system implements this trait as
+/// well.
+///
+/// # Examples
+///
+/// ```
+/// use bevy_ecs::prelude::*;
+///
+/// fn my_system_function(a_usize_local: Local<usize>) {}
+///
+/// let system = IntoSystem::into_system(my_system_function);
+/// ```
+// This trait has to be generic because we have potentially overlapping impls, in particular
+// because Rust thinks a type could impl multiple different `FnMut` combinations
+// even though none can currently
+pub trait IntoSystem<In, Out, Marker>: Sized {
+    type System: System<In = In, Out = Out>;
+    /// Turns this value into its corresponding [`System`].
+    fn into_system(this: Self) -> Self::System;
+
+    /// Pass the output of this system `A` into a second system `B`, creating a new compound system.
+    ///
+    /// The second system must have `In<T>` as its first parameter, where `T`
+    /// is the return type of the first system.
+    fn pipe<B, Final, MarkerB>(self, system: B) -> PipeSystem<Self::System, B::System>
+    where
+        B: IntoSystem<Out, Final, MarkerB>,
+    {
+        let system_a = IntoSystem::into_system(self);
+        let system_b = IntoSystem::into_system(system);
+        let name = format!("Pipe({}, {})", system_a.name(), system_b.name());
+        PipeSystem::new(system_a, system_b, Cow::Owned(name))
+    }
+}
+
+// Systems implicitly implement IntoSystem
+impl<In, Out, Sys: System<In = In, Out = Out>> IntoSystem<In, Out, ()> for Sys {
+    type System = Sys;
+    fn into_system(this: Self) -> Sys {
+        this
+    }
 }
 
 /// A collection of common adapters for [piping](super::PipeSystem) the result of a system.
